fitin/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.14
distro==1.9.0
Django==5.1.3
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
netaddr==1.3.0
pillow==11.0.0
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yookassa==3.4.1
